# ggml : remove bit shuffling