Implementation of #1241

Avoid unnecessary bit shuffling by packing the quants in a better way.
Requires model re-quantization

- `Q4_0`
  - [x] quantize
  - [x] dequantize
  - [x] dot ARM NEON
  - [x] dot AVX

- `Q4_1`
  - [x] quantize
  - [x] dequantize
  - [x] dot ARM NEON
  - [x] dot AVX

- `Q5_0`
  - [x] quantize
  - [x] dequantize
  - [x] dot ARM NEON
  - [x] dot AVX
  - [x] dot WASM SIMD

- `Q5_1`
  - [x] quantize
  - [x] dequantize
  - [x] dot ARM NEON
  - [x] dot AVX
  - [x] dot WASM SIMD

---

New timings:

| Model | Measure      | F16    | Q4_0   | Q4_1   | Q5_0   | Q5_1   | Q8_0   |
|------:|--------------|-------:|-------:|-------:|-------:|-------:|-------:|
|    7B | ms/tok @ 4th |    127 |     49 |     56 |     89 |     92 |     74 |
|    7B | ms/tok @ 8th |    120 |     44 |     52 |     49 |     52 |     70 |
|   13B | ms/tok @ 4th |    261* |    91 |    103 |    173 |    177 |    139 |
|   13B | ms/tok @ 8th |    316* |     81 |     95 |    103 |    113 |    134 |

* *these numbers vary a lot since the model is on the 32GB limit of my MacBook*

Old timings:

| Model | Measure      | F16    | Q4_0   | Q4_1   | Q5_0   | Q5_1   | Q8_0   |
|------:|--------------|-------:|-------:|-------:|-------:|-------:|-------:|
|    7B | ms/tok @ 4th |    128 |     56 |     61 |     91 |     95 |     75 |
|    7B | ms/tok @ 8th |    128 |     47 |     55 |     53 |     59 |     75 |
|   13B | ms/tok @ 4th |    239 |    104 |    113 |    176 |    185 |    141 |
|   13B | ms/tok @ 8th |    240 |     85 |     99 |    108 |    117 |    147 |

*overall, all these numbers seem to have about +/- 10% variablility from run to run. not ideal benchmark, but not sure what else to do*