@ggerganov 2023-05-06
@sw 

> I still think this could be done in a non-breaking manner for Q4, by simply changing the order of Q8

Yes, I'm still hesitating. But I think `Q8` quantization will be sub-optimal this way.
Not much, but still.