@ggerganov 2023-05-07
> Q5 interleaves the nibbles but not the single MSBs, that could be confusing. The scalar implementation seems to be broken, maybe it's because of that?

Somehow perplexity computation with `Q5_0` + cuBLAS is currently broken and I don't see the error.
It works on M1.