@ggerganov 2023-05-08
Surprising that `Q5_0` and `Q5_1` with 4 threads is slower than `Q8_0`