@eiery 2023-05-08
To lessen the breaking impact of this change (I got sent here by the big banner on the front page!) a conversion script could be provided for existing models. Since the weights themselves remain unchanged it should be possible for a script to repack them in the correct format.