@digiwombat 2023-05-08
I was somewhat waiting to see if anyone would raise this idea.

Is there enough bloat caused by supporting the old models that a header change to allow loading of both the new and old versions would be an imposition? That would keep old model compatibility and allow new quants to benefit from the speedup without any impact on end users.

Alternatively, if @eiery's proposal is feasible, detecting the old versions and giving an error with a copy/paste-able command would be very valuable. Or just a flag allowing it to do an in-place swap. This would definitely help front-end users for things like ooba's webui and koboldcpp.