@LostRuins 2023-05-09
Yes, if this is going to be a complete non-backwards compatible change, might I humbly suggest changing the magic in the file header at least, or utilizing the version field?