@Green-Sky 2023-05-09
https://github.com/ggerganov/llama.cpp/blob/0e48eb6f6b3588d78656267b3b8029b7711f6cdf/ggml.h#L191

can we increment this value by 1 ?