@Green-Sky 2023-05-15
@ProfessorSparrs if you have the f16 files, qnantizing is very easy and WAY less recource intensive than running the model. :) (check the `quantize` executable)